# gridsphere-backend/src/config/openapi.yaml
openapi: 3.0.3
info:
  title: GridSphere Weather Station API
  description: >
    Enterprise IoT backend migrated from FastAPI to Node.js/Express/TypeScript/Prisma.
    All routes, request/response shapes, and status codes match the original API exactly.
  version: "2.0.0"
servers:
  - url: /
    description: Current server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Paste the `access_token` returned by POST /login. Click "Authorize"
        above and enter just the token (Swagger UI adds the "Bearer " prefix).
  schemas:
...
    DeviceSensorCreate:
      type: object
      required: [device_id, sensor_type_id, sensor_label]
      properties:
        device_id: { type: integer, example: 1 }
        sensor_type_id: { type: integer, example: 1 }
        sensor_label: { type: string, example: "temp" }
        hardware_port: { type: string, nullable: true, example: "A0" }
        calibration_offset: { type: number, default: 0.0 }
        calibration_scale: { type: number, default: 1.0 }
...
    ApiError:
      type: object
      properties:
        detail: { type: string, example: "Could not validate credentials" }
paths:
...
  /sensors/device:
    post:
      tags: [Sensors]
      summary: Install a new sensor on a device
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DeviceSensorCreate' }
      responses:
        "201":
          description: Sensor installed
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: "success" }
                  data: { type: object }
        "404":
          description: Device not found/unauthorized or Sensor Type not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
...